requirements.txt
Flask
numpy
pillow
requests
easyocr
scikit-learn
gunicorn
#.venv\Scripts\activate